regroup all scheduling tests in the same file
